requirements.txt
python-dotenv
azure-core
azure-search-documents
openai
streamlit
